Revert creating ProcessedDocValues.java
