+++ server/src/main/java/org/elasticsearch/index/fielddata/HistogramValues.java
@@ -9,12 +9,15 @@
 
 package org.elasticsearch.index.fielddata;
 
+import org.apache.lucene.search.DocIdSetIterator;
+import org.elasticsearch.core.Nullable;
+
 import java.io.IOException;
 
 /**
  * Per-segment histogram values.
  */
-public abstract class HistogramValues implements ProcessedDocValues {
+public abstract class HistogramValues {
 
     /**
      * Get the {@link HistogramValue} associated with the current document.
@@ -22,8 +25,22 @@ public abstract class HistogramValues implements ProcessedDocValues {
      */
     public abstract HistogramValue histogram() throws IOException;
 
-    @Override
+    /** Advance the iterator to exactly {@code target} and return whether
+     *  {@code target} has a value.
+     *  {@code target} must be greater than or equal to the current
+     *  doc ID and must be a valid doc ID, ie. &ge; 0 and
+     *  &lt; {@code maxDoc}.*/
+    public abstract boolean advanceExact(int target) throws IOException;
+
     public int docValueCount() {
         return 1;
+    }
+
+    /**
+     * @return a doc id iterator over the doc values when available, otherwise null.
+     */
+    @Nullable
+    public DocIdSetIterator docIdIterator() {
+        return null;
     }
 }
