--- server/src/main/java/org/elasticsearch/index/fielddata/ProcessedDocValues.java
@@ -1,44 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-package org.elasticsearch.index.fielddata;
-
-import org.apache.lucene.search.DocIdSetIterator;
-import org.elasticsearch.core.Nullable;
-
-import java.io.IOException;
-
-/**
- * Interface that provides shared functionality for the elasticsearch doc values classes.
- */
-public interface ProcessedDocValues {
-
-    /** Advance the iterator to exactly {@code target} and return whether
-     *  {@code target} has a value.
-     *  {@code target} must be greater than or equal to the current
-     *  doc ID and must be a valid doc ID, ie. &ge; 0 and
-     *  &lt; {@code maxDoc}.*/
-    boolean advanceExact(int target) throws IOException;
-
-    /**
-     * Retrieves the number of values for the current document.  This must always
-     * be greater than zero.
-     * It is illegal to call this method after {@link #advanceExact(int)}
-     * returned {@code false}.
-     */
-    int docValueCount();
-
-    /**
-     * @return a doc id iterator over the doc values when available, otherwise null.
-     */
-    @Nullable
-    default DocIdSetIterator docIdIterator() {
-        return null;
-    }
-}
