+++ server/src/main/java/org/elasticsearch/index/fielddata/SortedNumericLongValues.java
@@ -18,13 +18,14 @@ import org.apache.lucene.index.PointValues;
 import org.apache.lucene.index.SortedNumericDocValues;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.search.LongValues;
+import org.elasticsearch.core.Nullable;
 
 import java.io.IOException;
 
 /**
  * Clone of {@link SortedNumericDocValues} for long values.
  */
-public abstract class SortedNumericLongValues implements ProcessedDocValues {
+public abstract class SortedNumericLongValues {
 
     private final boolean isSingleton;
     private final DocIdSetIterator docIdSetIterator;
@@ -59,6 +60,21 @@ public abstract class SortedNumericLongValues implements ProcessedDocValues {
         }
     };
 
+    /** Advance the iterator to exactly {@code target} and return whether
+     *  {@code target} has a value.
+     *  {@code target} must be greater than or equal to the current
+     *  doc ID and must be a valid doc ID, ie. &ge; 0 and
+     *  &lt; {@code maxDoc}.*/
+    public abstract boolean advanceExact(int target) throws IOException;
+
+    /**
+     * Retrieves the number of values for the current document.  This must always
+     * be greater than zero.
+     * It is illegal to call this method after {@link #advanceExact(int)}
+     * returned {@code false}.
+     */
+    public abstract int docValueCount();
+
     /**
      * Iterates to the next value in the current document. Do not call this more than
      * {@link #docValueCount} times for the document.
@@ -69,7 +85,7 @@ public abstract class SortedNumericLongValues implements ProcessedDocValues {
         return isSingleton;
     }
 
-    @Override
+    @Nullable
     public DocIdSetIterator docIdIterator() {
         return docIdSetIterator;
     }
